Weather_App/requirements.txt
pandas
numpy
requests
matplotlib
seaborn
streamlit
plotly
python-dotenv
scikit-learn